Refactor MCP tools
